自機と敵の挙動合わせ

+++ Assets/Imamoto/Scripts/UnitParam.cs
@@ -7,9 +7,9 @@ public struct UnitParam
     // 体力
     public int hp;
     // 攻撃力
-    public float attackPoint;
+    public int attackPoint;
     // 防御力
-    public float defensePoint;
+    public int defensePoint;
     // 移動速度
     public float moveSpeed;
     // 攻撃速度
